move robot pose and polygon to (0,0)
In GZ Ignition, tb3 publishes its ground truth pose as odometry with respect to its start position. Even when spawning the model at (5,5), the map->base_link transform is 0 because the robot did not drive.
This causes navigation to fail as the robot pose is too far from the start of the path.