Update to Jazzy/Rolling with F2C v1.2.1 (#82)
* [coverage_planning] make it build on rolling with F2C v1.2.1

* fix Werror=maybe-uninitialized

* restore tests and add missing ament_index_cpp to CMakeLists

* update demo launcher for Gazebo Ignition

* explicitly add models to avoid errors

* move robot pose and polygon to (0,0)

In GZ Ignition, tb3 publishes its ground truth pose as odometry with respect to its start position. Even when spawning the model at (5,5), the map->base_link transform is 0 because the robot did not drive.
This causes navigation to fail as the robot pose is too far from the start of the path.

* Add .gitignore

Copied from ros-navigation/navigation2

* Restore robot pose  & polygon to (5,5)

and add (5,5) to map->odom tf as it is the transform from world/map frame to the robot's initial pose

* Update row_coverage_demo_launch.py to Gazebo Ignition

* Note on F2C version

* Update test workflow

* lint and format

* add BTCPP_format to BT xmls

* CI: change F2C branch and update to Jazzy & Noble

* try F2C v1.2.1

* [tmp] ignore F2C compile errors

* Update lint step to jazzy

* test with cartesian polygon

* Revert "try F2C v1.2.1"

This reverts commit cb659870b7755b7150bc169d5d2018e50ccaa507.

* Revert "[tmp] ignore F2C compile errors"

This reverts commit 8164091dbb7292deeae9998a1ddf8a7f4ec8b10c.

* [tmp] rm failing test in opennav_coverage_bt

* switch test to cartesian_frame

* restore non-cartesian field test

* Revert "[tmp] rm failing test in opennav_coverage_bt"

This reverts commit 6a6194078fc636b8dd4446393bfb5da055483cde.

* restore rclcpp:shutdown()

Without the shutdown: The tests are passing in `RUN_ALL_TESTS()`, the test itself is failing with timeout. Most probably because `server_thread.join()` hangs forever.